+++ src/algorithms/changebonds/randexpand.jl
@@ -56,7 +56,7 @@ changebonds(ψ::MultilineMPS, alg::RandExpand) = changebonds!(copy(ψ), alg)
 
 function changebonds(ψ, H, alg::RandExpand, envs)
     newψ = changebonds(ψ, alg)
-    return newψ, environments(newψ, H)
+    return newψ, environments(newψ, H, newψ)
 end