+++ src/environments/multiple_envs.jl
@@ -10,31 +10,38 @@ Base.iterate(x::MultipleEnvironments) = iterate(x.envs)
 Base.iterate(x::MultipleEnvironments, i) = iterate(x.envs, i)
 
 # we need constructor, agnostic of particular MPS
-function environments(state, H::LazySum; kwargs...)
-    return MultipleEnvironments(map(Base.Fix1(environments, state), H.ops))
+function environments(below, H::LazySum, above = below; kwargs...)
+    return MultipleEnvironments(map(x -> environments(below, x, above; kwargs...), H.ops))
+end
+function environments(below, H::LazySum, above, alg; kwargs...)
+    return MultipleEnvironments(map(x -> environments(below, x, above, alg; kwargs...), H.ops))
 end
 
 function environments(
-        st::WindowMPS, H::LazySum;
+        st::WindowMPS, H::LazySum, above = st;
         lenvs = environments(st.left_gs, H), renvs = environments(st.right_gs, H)
     )
     return MultipleEnvironments(
         map(
-            (op, sublenv, subrenv) -> environments(st, op; lenvs = sublenv, renvs = subrenv),
+            (op, sublenv, subrenv) -> environments(st, op, above; lenvs = sublenv, renvs = subrenv),
             H.ops, lenvs, renvs
         )
     )
 end
 
-# we need to define how to recalculate
-"""
-    Recalculate in-place each sub-env in MultipleEnvironments
-"""
 function recalculate!(
         envs::MultipleEnvironments, below, operator::LazySum, above = below; kwargs...
     )
     for (subenvs, subO) in zip(envs.envs, operator)
         recalculate!(subenvs, below, subO, above; kwargs...)
+    end
+    return envs
+end
+function recalculate!(
+        envs::MultipleEnvironments, below, operator::LazySum, above, alg; kwargs...
+    )
+    for (subenvs, subO) in zip(envs.envs, operator)
+        recalculate!(subenvs, below, subO, above, alg; kwargs...)
     end
     return envs
 end